Faruk's final commit
