+++ .gitignore
@@ -163,3 +163,4 @@ data/
 *.db
 *.sqlite
 *.sqlite3
+logs
